+++ History.rst
@@ -13,6 +13,16 @@
    limitations under the License.
 
 
+**Version 0.44.0 (29.05.2026)**
+
+    *Added:*
+
+    * Search function
+
+    *Maintenance:*
+
+    * LaTeX preamble (``xcolor`` fix)
+
 **Version 0.43.0 (18.05.2026)**
 
     *Maintenance:*